backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
groq==0.9.0
httpx==0.27.0
redis
python-dotenv
pydantic


